Get pinned prompts into the input bar
